Separate resizing module from utils
